# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bookshelf - Wattpad-style book sharing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get or create current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a public user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Books ─────────────────────────────────────────────────────────
...
  /library:
    get:
      operationId: getMyLibrary
      tags: [library]
      summary: Get current user's saved library
      responses:
        "200":
          description: Library entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /library/{bookId}:
    post:
      operationId: addToLibrary
      tags: [library]
      summary: Add a book to personal library
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added to library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        "400":
          description: Already in library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      operationId: removeFromLibrary
      tags: [library]
      summary: Remove a book from personal library
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not in library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Comments ──────────────────────────────────────────────────────
...
  /messages/conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List all conversations for current user
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /messages/{otherUserId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation with another user
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a direct message to another user
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Storage ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Users ──────────────────────────────────────
    UserProfile:
      type: object
      required: [id, clerkId, username, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bookCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarObjectPath:
          type: string

    # ── Books ──────────────────────────────────────
    Book:
      type: object
      required: [id, title, authorId, authorName, fileObjectPath, fileFormat, isPublic, downloadCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        authorId:
          type: string
        authorName:
          type: string
        coverObjectPath:
          type: ["string", "null"]
        fileObjectPath:
          type: string
        fileFormat:
          type: string
          enum: [pdf, epub]
        isPublic:
          type: boolean
        downloadCount:
          type: integer
        commentCount:
          type: integer
        inLibrary:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LibraryEntry:
      type: object
      required: [id, bookId, userId, savedAt, book]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: string
        savedAt:
          type: string
          format: date-time
        book:
          $ref: "#/components/schemas/Book"
    LibraryListResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LibraryEntry"

    # ── Comments ───────────────────────────────────
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageListResponse:
      type: object
      required: [messages, otherUser]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        otherUser:
          $ref: "#/components/schemas/UserProfile"
    Conversation:
      type: object
      required: [otherUser, lastMessage, unreadCount]
      properties:
        otherUser:
          $ref: "#/components/schemas/UserProfile"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
    ConversationListResponse:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
    SendMessageBody:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Storage ────────────────────────────────────
...